Update dependencies as far as MSRV 1.85 allows

+++ src/device.rs
@@ -12,6 +12,7 @@ use crate::{
 
 bitflags! {
     /// Network device flags.
+    #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
     pub struct IfFlags: u32 {
         /// Set if the device is a loopback interface
         const LOOPBACK = raw::PCAP_IF_LOOPBACK;
@@ -266,7 +267,7 @@ impl Address {
             return None;
         }
 
-        match (*ptr).sa_family as u32 {
+        match (*ptr).sa_family {
             WinSock::AF_INET => {
                 let ptr: *const WinSock::SOCKADDR_IN = std::mem::transmute(ptr);
                 let addr: [u8; 4] = ((*ptr).sin_addr.S_un.S_addr).to_ne_bytes();
@@ -371,8 +372,7 @@ mod tests {
     impl InetAddressV4 for WinSock::SOCKADDR_IN {
         fn new() -> Self {
             let mut addr: Self = unsafe { std::mem::zeroed() };
-            // The cast is only necessary due to a bug in windows_sys@v0.36.1
-            addr.sin_family = WinSock::AF_INET as u16;
+            addr.sin_family = WinSock::AF_INET;
             addr
         }
 
@@ -417,8 +417,7 @@ mod tests {
     impl InetAddressV6 for WinSock::SOCKADDR_IN6 {
         fn new() -> Self {
             let mut addr: Self = unsafe { std::mem::zeroed() };
-            // The cast is only necessary due to a bug in windows_sys@v0.36.1
-            addr.sin6_family = WinSock::AF_INET6 as u16;
+            addr.sin6_family = WinSock::AF_INET6;
             unsafe {
                 addr.sin6_addr.u.Byte[0] = 0xFE;
                 addr.sin6_addr.u.Byte[1] = 0x80;
